requirements.txt
dj-database-url==2.0.0
Django==4.2.2
psycopg2-binary==2.9.6
gunicorn==20.1.0
djangorestframework==3.14.0